objectstore idempotent

+++ web/app/datasets/imports/download_datafiles.py
@@ -9,20 +9,21 @@ delivery of new data for this project.)
 """
 import os
 import logging
-import subprocess
-
+# import subprocess
+from pathlib import Path
 import objectstore
 
 logging.getLogger('requests').setLevel(logging.WARNING)
 logging.getLogger('urllib3').setLevel(logging.WARNING)
 logging.getLogger('swiftclient').setLevel(logging.WARNING)
 
-logger = logging.getLogger(__name__)
+log = logging.getLogger(__name__)
 
 FORMAT = '%(asctime)-15s %(message)s'
 logging.basicConfig(format=FORMAT, level=logging.DEBUG)
 
-GASTRANSITIE_OBJECTSTORE_PASSWORD = os.environ['GASTRANSITIE_OBJECTSTORE_PASSWORD']
+GASTRANSITIE_OBJECTSTORE_PASSWORD = \
+    os.environ['GASTRANSITIE_OBJECTSTORE_PASSWORD']
 
 OBJECTSTORE = {
     'VERSION': '2.0',
@@ -47,16 +48,35 @@ DATASETS = set([
 ])
 
 
+def file_exists(target):
+    target = Path(target)
+    return target.is_file()
+
+
 def download_container(conn, container, datadir):
-    logger.debug('Downloading dataset: %s', container['name'])
+
+    log.debug('Downloading dataset: %s', container['name'])
+
     target_dir = os.path.join(datadir, container['name'])
     os.makedirs(target_dir, exist_ok=True)
 
     content = objectstore.get_full_container_list(conn, container['name'])
-    for obj in content:
-        target_filename = os.path.join(target_dir, obj['name'])
+
+    for obj_meta in content:
+
+        target_filename = os.path.join(
+                target_dir, obj_meta['name'])
+
+        if file_exists(target_filename):
+            # Already downloaded
+            log.debug('Already downloaded %s', target_filename)
+            continue
+
+        # TODO age check!
+
         with open(target_filename, 'wb') as new_file:
-            _, obj_content = conn.get_object(container['name'], obj['name'])
+            _, obj_content = conn.get_object(
+                container['name'], obj_meta['name'])
             new_file.write(obj_content)
 
 
@@ -70,26 +90,14 @@ def download_containers(conn, datasets, datadir):
     * assumes we are running in a clean container (i.e. empty local data dir)
     * do not overwrite / delete old data
     """
-    logger.debug('Checking local data directory exists and is empty')
+    log.debug('Checking local data directory exists')
     if not os.path.exists(datadir):
         raise Exception('Local data directory does not exist.')
 
-    logger.debug('cache dir ls -l output:')
-    p = subprocess.Popen(['ls', '-l', datadir])
-    p.wait()
-
-    listing = os.listdir(datadir)
-    if listing:
-        if len(listing) == 1 and listing[0] == 'README':
-            # Case where the 'data' dictory is used from a fresh checkout.
-            pass
-        else:
-            raise Exception('Local data directory not empty!')
-
-    logger.debug('Establishing object store connection.')
+    log.debug('Establishing object store connection.')
     resp_headers, containers = conn.get_account()
 
-    logger.debug('Downloading containers ...')
+    log.debug('Downloading containers ...')
     for c in containers:
         if c['name'] in datasets:
             download_container(conn, c, datadir)
